+++ changelogs/fragments/windows-exec-changes.yaml
@@ -1,0 +1,2 @@
+minor_changes:
+- include better error handling for Windows errors to help with debugging module errors
